backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
motor
PyJWT
passlib[bcrypt]
httpx
scikit-learn
pandas
numpy
